+++ Engine/Manager/TextureManager.cpp
@@ -79,7 +79,7 @@ void TextureManager::LoadTexture(const std::string& filePath) {
 	textureData.srvIndex = srvManager_->Allocate();
 	textureData.srvHandleCPU = srvManager_->GetCPUDescriptorHandle(textureData.srvIndex);
 	textureData.srvHandleGPU = srvManager_->GetGPUDescriptorHandle(textureData.srvIndex);
-	srvManager_->CreateSRVforTexture2D(textureData.srvIndex, textureData.resource.Get(), textureData.metadata.format, textureData.metadata.mipLevels, textureData.metadata);
+	srvManager_->CreateSRVforTexture2D(textureData.srvIndex, textureData.resource.Get(), textureData.metadata.format, (UINT)textureData.metadata.mipLevels, textureData.metadata);
 }
 
 void TextureManager::LoadTexture(const std::string& directoryPath, const std::string& fileName) {
@@ -134,7 +134,7 @@ void TextureManager::LoadTexture(const std::string& directoryPath, const std::st
 	textureData.srvIndex = srvManager_->Allocate();
 	textureData.srvHandleCPU = srvManager_->GetCPUDescriptorHandle(textureData.srvIndex);
 	textureData.srvHandleGPU = srvManager_->GetGPUDescriptorHandle(textureData.srvIndex);
-	srvManager_->CreateSRVforTexture2D(textureData.srvIndex, textureData.resource.Get(), textureData.metadata.format, textureData.metadata.mipLevels, textureData.metadata);
+	srvManager_->CreateSRVforTexture2D(textureData.srvIndex, textureData.resource.Get(), textureData.metadata.format, (UINT)textureData.metadata.mipLevels, textureData.metadata);
 }
 
 uint32_t TextureManager::GetTextureIndexByFilePath(const std::string& filePath) {
